requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
django-jazzmin==2.6.0
djangorestframework==3.16.1
drf-extra-fields==3.7.0
drf-yasg==1.21.11
filetype==1.2.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
